Move `Product-` and `SumDerivationMethods` inside `Product-`/`SumDerivation` companions
